requirement.txt
asgiref==3.8.1
Django==4.2.11
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
PyJWT==2.10.1
python-dotenv==1.0.1
pytz==2025.2
sqlparse==0.5.3
